beheer onderdelen afgewerkt

+++ Autohandel.web/Areas/Admin/Controllers/CategorieOnderdelenController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using Autohandel.Domain.Data;
 using Autohandel.Domain.Entities;
+using Autohandel.web.ViewModels.CategorieOnderdelenViewModels;
 
 namespace Autohandel.web.Areas.Admin.Controllers
 {
@@ -23,7 +24,12 @@ namespace Autohandel.web.Areas.Admin.Controllers
         // GET: Admin/CategorieOnderdelen
         public async Task<IActionResult> Index()
         {
-            var autohandelContext = _context.CategorieOnderdelen.Include(c => c.Parent);
+            var autohandelContext = _context.CategorieOnderdelen
+                                    .Include(c => c.Parent)
+                                    .Include(p => p.Products)
+                                    .OrderBy(c => c.Parent.OnderdelenCategorienaam)
+                                    .ThenBy(ch => ch.OnderdelenCategorienaam)
+                ;
             return View(await autohandelContext.ToListAsync());
         }
 
@@ -46,11 +52,40 @@ namespace Autohandel.web.Areas.Admin.Controllers
             return View(categorieOnderdelen);
         }
 
+        private List<SelectListItem> hoofdcategorieOnderdelenlijst()
+        {
+            var list =
+            //return
+                    _context.CategorieOnderdelen
+                     .Include(c => c.Parent)
+                     .Where(hc => hc.ParentId == null)
+                .Select(hc => new SelectListItem
+                { //transformeer elke hoofdcategorie naar een SelectListItem
+                    Value = hc.OnderdelenCategorieId.ToString(),
+                    Text = hc.OnderdelenCategorienaam
+
+                })
+                .ToList();
+            list.Add(new SelectListItem { Value = null, Text = "", Selected = true });
+            return list;
+        }
+
+
         // GET: Admin/CategorieOnderdelen/Create
         public IActionResult Create()
         {
-            ViewData["ParentId"] = new SelectList(_context.CategorieOnderdelen, "OnderdelenCategorieId", "OnderdelenCategorienaam");
-            return View();
+            //ViewData["ParentId"] = new SelectList(_context.CategorieOnderdelen, "OnderdelenCategorieId", "OnderdelenCategorienaam");
+            
+            var vm = new CategorieOnderdelenCreateViewModel()
+            {
+
+                CategorieOnderdeel = null,
+                HoofdcategorieOnderdelenlijst = hoofdcategorieOnderdelenlijst()
+            };
+
+
+            return View(vm);
+
         }
 
         // POST: Admin/CategorieOnderdelen/Create
@@ -58,16 +93,63 @@ namespace Autohandel.web.Areas.Admin.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("OnderdelenCategorieId,OnderdelenCategorienaam,ParentId")] CategorieOnderdelen categorieOnderdelen)
+        public async Task<IActionResult> Create([Bind("OnderdelenCategorieId,OnderdelenCategorienaam,ParentId, CategorieOnderdeel, HoofdcategorieOnderdelenlijst ")] CategorieOnderdelenCreateViewModel inputCatOndCreateVm)
         {
+                      
+            if (CategorieNaamOnderdelenExists(inputCatOndCreateVm.CategorieOnderdeel.OnderdelenCategorienaam))
+            {
+                ModelState.AddModelError("NaamValidatieError", "Deze categorie is reeds aangemaakt!");
+            }
+
             if (ModelState.IsValid)
             {
-                _context.Add(categorieOnderdelen);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                //    _context.Add(categorieOnderdelen);
+                //    await _context.SaveChangesAsync();
+                //    return RedirectToAction(nameof(Index));
+                //}
+                //ViewData["ParentId"] = new SelectList(_context.CategorieOnderdelen, "OnderdelenCategorieId", "OnderdelenCategorienaam", categorieOnderdelen.ParentId);
+                //return View(categorieOnderdelen);
+
+
+
+                if (inputCatOndCreateVm.CategorieOnderdeel.OnderdelenCategorienaam != null) //zo ja,
+                {
+                    //nieuwe categorie maken
+                    var categorieOnderdeelToAdd = new CategorieOnderdelen
+                    {
+                        OnderdelenCategorienaam = inputCatOndCreateVm.CategorieOnderdeel.OnderdelenCategorienaam,
+                        ParentId = inputCatOndCreateVm.CategorieOnderdeel.ParentId,
+
+                    };
+
+                    //categorie tovoegen aan de dbSet (tabel)
+                    _context.CategorieOnderdelen.Add(categorieOnderdeelToAdd);
+
+                    //context wijzigingen doorvoeren naar de Database
+                    await _context.SaveChangesAsync();
+                    //actie voor response ondernemen
+                    TempData["SuccessMessage"] = $"De categorie <b>{categorieOnderdeelToAdd.OnderdelenCategorienaam}</b> werd toegevoegd!";
+                    return RedirectToAction(nameof(Index));
+                }
             }
-            ViewData["ParentId"] = new SelectList(_context.CategorieOnderdelen, "OnderdelenCategorieId", "OnderdelenCategorienaam", categorieOnderdelen.ParentId);
-            return View(categorieOnderdelen);
+            else
+            {
+                //de existing categorie bestaat niet
+                ModelState.AddModelError("CategorieOnderdeel.ParentId",
+                    $"Input is niet correct!");
+
+
+            }
+            //model not valid
+
+            //input model wordt nu het view model, dus moet nog vervolledigd worden
+            inputCatOndCreateVm = new CategorieOnderdelenCreateViewModel()
+            {
+                CategorieOnderdeel = null,
+                HoofdcategorieOnderdelenlijst = hoofdcategorieOnderdelenlijst()
+            };
+
+            return View(inputCatOndCreateVm);
         }
 
         // GET: Admin/CategorieOnderdelen/Edit/5
@@ -78,13 +160,55 @@ namespace Autohandel.web.Areas.Admin.Controllers
                 return NotFound();
             }
 
-            var categorieOnderdelen = await _context.CategorieOnderdelen.SingleOrDefaultAsync(m => m.OnderdelenCategorieId == id);
+           
+
+            var categorieOnderdelen = await _context.CategorieOnderdelen
+                                            .Include(c => c.Children)
+                                            .Include(p => p.Products)
+                                            .SingleOrDefaultAsync(m => m.OnderdelenCategorieId == id);
             if (categorieOnderdelen == null)
             {
                 return NotFound();
             }
-            ViewData["ParentId"] = new SelectList(_context.CategorieOnderdelen.Where(hc => hc.ParentId == null), "OnderdelenCategorieId", "OnderdelenCategorienaam", categorieOnderdelen.ParentId);
-            return View(categorieOnderdelen);
+            if ((categorieOnderdelen.Products.Count() != 0) || (categorieOnderdelen.Children.Count() != 0))
+            {
+                try
+                {
+                    TempData["AlertMessage"] = $"De categorie <b>{categorieOnderdelen.OnderdelenCategorienaam}</b> mag niet gewijzigd worden." +
+                        $"De categorienaam is een gekoppeld aan een (sub)categorie of is een categorie waaronder reeds producten zijn opgeslagen!";
+
+                    return RedirectToAction(nameof(Index));
+                }
+                catch
+                {
+                    TempData["AlertMessage"] = $"Systeemfout in categorieënbeheer!";
+
+                    return RedirectToAction(nameof(Index));
+                }
+            }
+            CategorieOnderdelenCreateViewModel vm;
+            if (categorieOnderdelen.ParentId == null)
+            {
+                 vm = new CategorieOnderdelenCreateViewModel()
+
+                {
+                    CategorieOnderdeel = categorieOnderdelen,
+                    HoofdcategorieOnderdelenlijst = null
+                };
+            }
+            else
+            {
+                 vm = new CategorieOnderdelenCreateViewModel()
+
+                {
+                    CategorieOnderdeel = categorieOnderdelen,
+                    HoofdcategorieOnderdelenlijst = hoofdcategorieOnderdelenlijst()
+                };
+            }
+
+            
+            //ViewData["ParentId"] = new SelectList(_context.CategorieOnderdelen.Where(hc => hc.ParentId == null), "OnderdelenCategorieId", "OnderdelenCategorienaam", categorieOnderdelen.ParentId);
+            return View(vm);
         }
 
         // POST: Admin/CategorieOnderdelen/Edit/5
@@ -92,9 +216,14 @@ namespace Autohandel.web.Areas.Admin.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("OnderdelenCategorieId,OnderdelenCategorienaam,ParentId")] CategorieOnderdelen categorieOnderdelen)
+        public async Task<IActionResult> Edit(int id, [Bind("OnderdelenCategorieId,OnderdelenCategorienaam,ParentId,CategorieOnderdeel, HoofdcategorieOnderdelenlijst")] CategorieOnderdelenCreateViewModel editCatOndCreateVm)
         {
-            if (id != categorieOnderdelen.OnderdelenCategorieId)
+            if (CategorieNaamOnderdelenMoreExists(id, editCatOndCreateVm.CategorieOnderdeel.OnderdelenCategorienaam))
+            {
+                ModelState.AddModelError("NaamValidatieError", "Dit is de naam van een bestaande categorie!");
+            }
+
+            if (id != editCatOndCreateVm.CategorieOnderdeel.OnderdelenCategorieId)
             {
                 return NotFound();
             }
@@ -103,24 +232,46 @@ namespace Autohandel.web.Areas.Admin.Controllers
             {
                 try
                 {
-                    _context.Update(categorieOnderdelen);
-                    await _context.SaveChangesAsync();
+
+                    if (editCatOndCreateVm.CategorieOnderdeel.OnderdelenCategorienaam != null) //zo ja,
+                    {
+                        //bestaande categorie wijzigen
+                        var categorieOnderdeelToUpdate = await _context.CategorieOnderdelen.SingleOrDefaultAsync(m => m.OnderdelenCategorieId == id);
+                        if (categorieOnderdeelToUpdate == null)
+                        {
+                            return NotFound();
+                        }
+                        categorieOnderdeelToUpdate.OnderdelenCategorienaam = editCatOndCreateVm.CategorieOnderdeel.OnderdelenCategorienaam;
+                        categorieOnderdeelToUpdate.ParentId = editCatOndCreateVm.CategorieOnderdeel.ParentId;
+                      
+                        //categorie updaten
+                        _context.CategorieOnderdelen.Update(categorieOnderdeelToUpdate);
+
+                        //context wijzigingen doorvoeren naar de Database
+                        await _context.SaveChangesAsync();
+                        //actie voor response ondernemen
+                        TempData["SuccessMessage"] = $"De categorie <b>{categorieOnderdeelToUpdate.OnderdelenCategorienaam}</b> werd gewijzigd!";
+                        return RedirectToAction(nameof(Index));
+                    }
+                   
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!CategorieOnderdelenExists(categorieOnderdelen.OnderdelenCategorieId))
+                    if (!CategorieOnderdelenExists(editCatOndCreateVm.CategorieOnderdeel.OnderdelenCategorieId))
                     {
                         return NotFound();
                     }
                     else
                     {
+                        ModelState.AddModelError("CategorieOnderdeel.ParentId",
+                        $"Input is niet correct!");
                         throw;
                     }
                 }
                 return RedirectToAction(nameof(Index));
             }
-            ViewData["ParentId"] = new SelectList(_context.CategorieOnderdelen, "OnderdelenCategorieId", "OnderdelenCategorienaam", categorieOnderdelen.ParentId);
-            return View(categorieOnderdelen);
+            //ViewData["ParentId"] = new SelectList(_context.CategorieOnderdelen, "OnderdelenCategorieId", "OnderdelenCategorienaam", categorieOnderdelen.ParentId);
+            return View(editCatOndCreateVm);
         }
 
         // GET: Admin/CategorieOnderdelen/Delete/5
@@ -132,11 +283,24 @@ namespace Autohandel.web.Areas.Admin.Controllers
             }
 
             var categorieOnderdelen = await _context.CategorieOnderdelen
-                .Include(c => c.Parent)
+                .Include(c => c.Children)
+                .Include(p => p.Products)
                 .SingleOrDefaultAsync(m => m.OnderdelenCategorieId == id);
             if (categorieOnderdelen == null)
             {
                 return NotFound();
+            }
+
+            if ((categorieOnderdelen.Children.Count() != 0)||(categorieOnderdelen.Products.Count() != 0))
+            {
+                try
+                {
+                    TempData["AlertMessage"] = $"De categorie <b>{categorieOnderdelen.OnderdelenCategorienaam}</b> mag niet verwijderd worden." +
+                        $"De categorienaam is reeds verbonden met subcategorie of producten.";
+                    
+                    return RedirectToAction(nameof(Index));
+                }
+                catch { }
             }
 
             return View(categorieOnderdelen);
@@ -147,15 +311,52 @@ namespace Autohandel.web.Areas.Admin.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var categorieOnderdelen = await _context.CategorieOnderdelen.SingleOrDefaultAsync(m => m.OnderdelenCategorieId == id);
-            _context.CategorieOnderdelen.Remove(categorieOnderdelen);
-            await _context.SaveChangesAsync();
+            var categorieOnderdelen = await _context.CategorieOnderdelen
+                                            .Include(c=>c.Children)  
+                                            .Include(p=>p.Products)
+                                            .SingleOrDefaultAsync(m => m.OnderdelenCategorieId == id);
+           
+            if (categorieOnderdelen.Products.Count() == 0)
+            {
+                try
+                {
+                    _context.CategorieOnderdelen.Remove(categorieOnderdelen);
+                    await _context.SaveChangesAsync();
+
+                    TempData["SuccessMessage"] = $"De categorie <b>{categorieOnderdelen.OnderdelenCategorienaam}</b> werd verwijderd!";
+                   
+                }
+                catch
+                {
+                    TempData["AlertMessage"] = $"De categorie <b>{categorieOnderdelen.OnderdelenCategorienaam}</b> kon niet verwijderd worden.";
+                   
+                }
+            }
+            else
+            {
+                TempData["AlertMessage"] = $"De categorie <b>{categorieOnderdelen.OnderdelenCategorienaam}</b> kon niet verwijderd worden. Er zijn reeds producten aan verbonden!";
+              
+
+            }
             return RedirectToAction(nameof(Index));
+
         }
 
         private bool CategorieOnderdelenExists(int id)
         {
             return _context.CategorieOnderdelen.Any(e => e.OnderdelenCategorieId == id);
         }
+
+        private bool CategorieNaamOnderdelenExists(string checkNaam)
+        {
+            return _context.CategorieOnderdelen.Any(c => c.OnderdelenCategorienaam == checkNaam); ;
+        }
+
+        private bool CategorieNaamOnderdelenMoreExists(int id,string checkNaam)
+        {
+            return _context.CategorieOnderdelen
+                              .Any(c => c.OnderdelenCategorienaam == checkNaam && c.OnderdelenCategorieId != id); ;
+        }
+
     }
 }
